LightsOn/requirements.txt
Flask>=2.2.5
google-cloud-tasks>=2.11.0
google-auth>=2.23.0
gunicorn>=20.1.0
